Move workflow into tasks module
