+++ src/lib.rs
@@ -8,9 +8,9 @@ pub(crate) mod rest;
 pub mod tasks;
 pub mod utils;
 pub mod worker;
-pub mod workflow;
 
 pub use client::HatchetClient;
 pub use error::HatchetError;
 pub use tasks::Context;
+pub use tasks::workflow::{TriggerWorkflowOptions, Workflow};
 pub use worker::Worker;
