+++ src/tasks/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod context;
 pub(crate) mod erased_task;
 pub mod task_trait;
+pub mod workflow;
 
 pub use context::Context;
 pub use erased_task::{ErasedTask, ErasedTaskFunction};
